SCAN4NET-333 Extract Incremental PR Analysis
